Exported function in index.js
