+++ lib/jsshift.js
@@ -642,7 +642,7 @@ function pluralizeReplaceWithTemplateString(j, path) {
 
 function transformer(file, api) {
   const j = api.jscodeshift;
-  let ast = j(file.source);
+  let ast = j(file.source ?? file);
   let somethingChanged = false;
 
   const replacedPluralize = ast
