+++ package.json
@@ -2,7 +2,6 @@
   "name": "@soxhub/intl-template-transform",
   "version": "1.1.0",
   "description": "",
-  "main": "runner.js",
   "bin": "./bin/cli.js",
   "scripts": {
     "start": "node runner.js",
